apps/agents/pyproject.toml
[project]
name = "ChurnQ-agents"
version = "0.1.0"
description = "Railway-hosted FastAPI service: Stripe webhooks, LangGraph churn prediction, CrewAI feedback, BullMQ jobs"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.115.0",
    "pydantic-settings>=2.6.0",
    "stripe>=11.0.0",
    "uvicorn[standard]>=0.32.0",
    "langgraph>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "langchain-core>=0.3.0",
    "scikit-learn>=1.5.0",
    "numpy>=1.26.0",
    "resend>=2.0.0",
    "apscheduler>=3.10.0",
    "voyageai>=0.3.0",
]

[project.scripts]
ChurnQ-agents = "churnq_agents.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/churnq_agents"]

[project.optional-dependencies]
dev = ["httpx>=0.27.0", "pytest>=8.0.0"]
